// src/generated/models/AirtimeRewardDTO.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type AirtimeRewardDTO = {
    id?: string;
    studentId?: string;
    phoneNumber?: string;
    network?: string;
    requestId?: string;
    amount?: number;
    requestedAt?: string;
    approved?: boolean;
    approvedAt?: string;
    message?: string;
    status?: string;
    balance?: number;
};